refactor: split the worker in multiple files
